Wrote Gomoku
This is "Freestyle Gomoku" generalized to be played on an infinite board.  On p. 60 of József Beck's book Combinatorial Games: Tic-Tac-Toe Theory, (2008) it says about "unrestricted 5-in-a-row" (or Gomoku on an infinite board): "Open Problem 4.1. Is it true that unrestricted 5-in-a-row is a first player win?"  I like to imagine someone spending a lifetime (an eternity?) trying every possible move in an attempt to find a win for white (seeing as this is an open problem at this time.)  

I think playing on an infinite board may actually SIMPLIFY things, code-wise, as one does not need to keep track of the size or edges of a board.  The code here simply keeps two list of coordinates played, one for each player, checks new moves to ensure they are not in either lists already, and then basically looks to the lists of the current player to see whether a newly placed move is now in part of a row of five moves in that player's move list in which case they win.